python script to generate qaoa n=6
Using Pennylane + qiskit, we  generate cicuits in the native gate set
directly, and we observe a discount of number of gates since we do not
need to transpile again.
To standardize the comparison, we should adopt the same qasm generation
for all qaoa circuits.

In QASMBenchmark, we have two sets of qaoa:
1. n=3 p=3
2. n=6 p=2
We adopt n=6 p=2 here and hopefully we will stick with p=2 for all qaoa
circuits.

OrderedDict([('cx', 36), ('h', 30), ('rz', 30), ('measure', 6)])